Handle feed burner xhtml atom feeds
